requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dependency-injector==4.41.0
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
filelock==3.15.4
gitdb==4.0.11
GitPython==3.1.41
idna==3.7
inflection==0.5.1
packaging==24.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
six==1.16.0
smmap==5.0.1
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
